django/requirements.txt
asgiref==3.10.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.10.5 ; python_version >= "3.10" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.10" and platform_python_implementation != "PyPy" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.10" and python_version < "4.0"
cryptography==46.0.3 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
django-superset-integration==0.1.17 ; python_version >= "3.10" and python_version < "4.0"
django==5.2.7 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.16.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.11 ; python_version >= "3.10" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.11 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.10" and platform_python_implementation != "PyPy" and python_version < "4.0" and implementation_name != "PyPy"
pyjwt==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2025.2 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
